requirements/apps.txt
django==1.3
South==0.7.3
gunicorn==0.12.2
python-dateutil==1.5
django-pagination==1.0.7
lxml==2.3
twilio==2.0.6
tropo-webapi-python==0.1.1
django-sorting==0.1
psycopg2==2.4
django-rosetta==0.6.0
django-celery==2.2.4
celery==2.2.7
python-twitter==0.8.2
feedparser==5.0.1
gevent==0.13.6
greenlet==0.3.1
distribute>=0.6.19
nose==1.0
django-selectable==0.5.0
-e git+https://github.com/caktus/rapidsms-decisiontree-app.git@develop#egg=decisiontree
-e git+https://github.com/dimagi/rapidsms-tropo.git@develop#egg=rtropo
-e git+https://github.com/caktus/rapidsms-threadless-router.git#egg=threadless_router
-e git+https://github.com/dimagi/rapidsms.git#egg=rapidsms
-e git+https://github.com/adammck/djtables#egg=djtables
couchdbkit==0.5.4
pytz
-e git+https://github.com/dimagi/couchlog.git#egg=couchlog
-e git+https://github.com/dimagi/auditcare.git#egg=auditcare
-e git+https://github.com/dimagi/dimagi-utils.git#egg=dimagi_utils
django-timezones==0.2
